requirements.txt
# Application dependencies
# Pin or relax versions as needed.
google-adk>=1.10.0
google-generativeai>=0.8.5
python-dotenv>=1.1.1
pymongo>=4.0.0
requests>=2.31.0
# Google Calendar / Google APIs
google-api-python-client>=2.129.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
